Project is fully finished!
All files have been cleaned up, removed OOD or useless code. Added a new `datetimeHandler` file which handles input for the `date` input fields within the forms. Projects are able to be created, with todos being added inside of them. Projects and todos have actions to delete a project or todo, with todos having an additional `edit` and `mark` action for editing and marking a todo as complete.